feat: added user's name dynamically after login
